fps optimizations and deleting unneded stuff

+++ project/data/terrain.frag
@@ -47,7 +47,7 @@ uniform vec3 fogColor;
 varying mediump float visibility;
 
 // World Space Coordinates
-varying highp vec3 v_normal;
+varying lowp vec3 v_normal;
 varying highp vec4 v_position;
 varying mediump vec3 v_tangent;
 varying mediump vec3 v_bitangent;
